Redirect 2.3.1 * Magento 2.3 compatibility * Refactor plugin according to PayGate PayWeb Plugin * OneStepCheckout compatibility
