+++ Redirect/SID/InstantEFT/composer.json
@@ -29,7 +29,7 @@
       "SID\\InstantEFT\\": ""
     }
   },
-  "version": "2.3.0",
+  "version": "2.3.1",
   "license": [
     "proprietary"
   ]
