+++ Redirect/SID/InstantEFT/view/frontend/web/template/payment/sid.html
@@ -30,7 +30,7 @@
             <!-- ko template: getTemplate() --><!-- /ko -->
             <!--/ko-->
         </div>
-        <div class="actions-toolbar">
+        <div class="actions-toolbar" style="display:block;">
             <div class="primary">
                 <button class="action primary checkout"
                         type="submit" data-bind="click: placeOrder,
